Add middleware support to thrift-client
