+++ packages/thrift-client/package.json
@@ -24,7 +24,7 @@
     "pretest": "npm run build:test",
     "test": "npm run test:only --",
     "test:watch": "watch 'npm run test:only' ./dist",
-    "test:only": "lab --timeout 10000 --verbose -l -S ./dist/**/*.spec.js"
+    "test:only": "lab --timeout 10000 --verbose -l -S -P spec dist/tests"
   },
   "author": "Credit Karma",
   "license": "Apache-2.0",
