+++ packages/thrift-client/src/main/connections/HttpConnection.ts
@@ -8,6 +8,7 @@ import {
 
 import {
   IHttpConnectionOptions,
+  IThriftMiddleware,
 } from './types'
 
 function normalizePath(path: string = '/'): string {
@@ -28,6 +29,7 @@ export abstract class HttpConnection<Context = never> implements IThriftConnecti
   protected hostName: string
   protected path: string
   protected protocol: HttpProtocol
+  protected middleware: IThriftMiddleware[]
 
   constructor(options: IHttpConnectionOptions) {
     this.port = options.port
@@ -36,11 +38,22 @@ export abstract class HttpConnection<Context = never> implements IThriftConnecti
     this.Transport = getTransport(options.transport)
     this.Protocol = getProtocol(options.protocol)
     this.protocol = ((options.https === true) ? 'https' : 'http')
+    this.middleware = []
   }
 
   public abstract write(dataToWrite: Buffer, context?: Context): Promise<Buffer>
 
+  public register(...middleware: Array<IThriftMiddleware>): void {
+    middleware.forEach((next: IThriftMiddleware) => {
+      this.middleware.push(next)
+    })
+  }
+
   public send(dataToSend: Buffer, context?: Context): Promise<Buffer> {
-    return this.write(dataToSend, context)
+    return this.write(dataToSend, context).then((data: Buffer) => {
+      return this.middleware.reduce((acc: Promise<Buffer>, next: IThriftMiddleware): Promise<Buffer> => {
+        return acc.then(next.handler)
+      }, Promise.resolve(data))
+    })
   }
 }
